requirements.txt
dnspython3
dnspython~=1.15.0